Establish baseline for chrome 25.0.1364.169.
Extract the release source from chrome repo using:
    gclient config https://src.chromium.org/chrome/releases/25.0.1364.169
    gclient sync --nohooks
    . build/android/envsetup.sh
    gclient runhooks

Then this is overlayed on the mips clank repo, tarball-469 branch (chrome
18.0.1025.469 tarball release).

This results in a very large commit, since there are lots of diffs.

This is take-2, since I did not clean all .git & .svn cruft from the
previous m25 release, resulting in some file corruption.

+++ WebKit/Source/WebKit/mac/History/WebHistoryItem.mm
@@ -47,7 +47,6 @@
 #import <WebCore/Image.h>
 #import <WebCore/KURL.h>
 #import <WebCore/PageCache.h>
-#import <WebCore/PlatformString.h>
 #import <WebCore/RunLoop.h>
 #import <WebCore/ThreadCheck.h>
 #import <WebCore/WebCoreObjCExtras.h>
@@ -55,6 +54,7 @@
 #import <wtf/Assertions.h>
 #import <wtf/MainThread.h>
 #import <wtf/StdLibExtras.h>
+#import <wtf/text/WTFString.h>
 
 // Private keys used in the WebHistoryItem's dictionary representation.
 // see 3245793 for explanation of "lastVisitedDate"
@@ -232,7 +232,7 @@ void WKNotifyHistoryItemChanged(HistoryItem*)
             [result appendString:[child description]];
         }
         // shift all the contents over.  A bit slow, but hey, this is for debugging.
-        NSRange replRange = {currPos, [result length]-currPos};
+        NSRange replRange = { static_cast<NSUInteger>(currPos), [result length] - currPos };
         [result replaceOccurrencesOfString:@"\n" withString:@"\n    " options:0 range:replRange];
     }
